requirements.txt
Pillow~=10.3.0
opencv-python~=4.9.0.80
numpy~=1.26.4
requests~=2.31.0
jieba~=0.42.1
pymongo
nonebot2[fastapi]
nonebot-adapter-onebot